Module jf_pcs::prelude

Prelude

Re-exports

pub use crate::errors::PCSError;

pub use crate::multilinear_kzg::MultilinearKzgBatchProof;

pub use crate::multilinear_kzg::MultilinearKzgPCS;

pub use crate::multilinear_kzg::MultilinearKzgProof;

pub use crate::multilinear_kzg::MLE;

pub use crate::univariate_kzg::UnivariateKzgBatchProof;

pub use crate::univariate_kzg::UnivariateKzgPCS;

pub use crate::univariate_kzg::UnivariateKzgProof;

pub use crate::PolynomialCommitmentScheme;

pub use crate::StructuredReferenceString;

Structs

Commitment

A commitment is an Affine point.

MultilinearProverParam

Prover Config

MultilinearUniversalParams

Universal Parameter

MultilinearVerifierParam

Verifier Config

UnivariateProverParam

UnivariateProverParam is used to generate a proof

UnivariateUniversalParams

UniversalParams are the universal parameters for the KZG10 scheme.

UnivariateVerifierParam

UnivariateVerifierParam is used to check evaluation proofs for a given commitment.

Functions

get_batched_nv

Return the number of variables that one need for an MLE to batch the list of MLEs

merge_polynomials

merge a set of polynomials. Returns an error if the polynomials do not share a same number of nvs.